First version of the full repository to create a virtual Hadoop Cluster.
It uses Vagrant to create the VirtualBox virtual machines and shell scripts for
the provisioning.

It comes with some test codes and documentation on how to run some basic tests
in the cluster so that we can test its proper setup.

In the future, we should change the provisioning to Ansible.